+++ Function testing/FontDataToStruct/F_ResizeStructs.c
@@ -6,23 +6,27 @@ int F_ResizeStructs(int FontSize, Characters Character){
 
     int i;
     int length;
-    int OldX;
-    int OldY;
-    int NewX;
-    int NewY;
+    float OldX;
+    float OldY;
+    float NewX;
+    float NewY;
 
     length = Character.length;
 
 
     for(i=0; i<length; i++){
         OldX = Character.Xpos[i];
-        NewX = OldX*((float)FontSize/(float)18);
+        //printf("\n\t\tOldX = %f",OldX);
+        NewX = (float)OldX*((float)FontSize/(float)18);
+        //printf("\n\t\tNewX = %f",NewX);
         Character.Xpos[i] = NewX;
     }
 
     for(i=0; i<length; i++){
         OldY = Character.Ypos[i];
-        NewY = OldY*((float)FontSize/(float)18);
+        //printf("\n\t\tOldY = %f",OldY);
+        NewY = (float)OldY*((float)FontSize/(float)18);
+        //printf("\n\t\tNewY = %f",NewY);
         Character.Ypos[i] = NewY;
     }
